refactor to return type: long

+++ src/main/java/com/epam/Fibonacci.java
@@ -2,7 +2,7 @@ package com.epam;
 
 public class Fibonacci {
 
-    public static int countFibonacciSequence(int elementNumber) {
+    public static long countFibonacciSequence(int elementNumber) {
         if (elementNumber < 0) {
             throw new IllegalArgumentException();
         }
